+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-# Tamr Terraform Template Repo - v0.1.0 - Feb 25th 2020
-* Initing project
+# TAMR AWS EMR Hbase root module - v0.1.0 - March 31st 2020
+
+### Added
+* Root module for EMR Hbase
+* Sub modules:
+    * S3 module
+    * Security Groups module
+    * IAM module
+    * EMR Hbase cluster module
+